backend/requirements.txt
fastapi
uvicorn
python-dotenv
openai
pydantic
chromadb
sentence-transformers
pypdf
